+++ Source/Plugins/NetworkActivityPlugin.swift
@@ -6,7 +6,7 @@ public enum NetworkActivityChangeType {
 }
 
 /// Provides each request with optional NSURLCredentials.
-public final class NetworkActivityPlugin: Plugin {
+public final class NetworkActivityPlugin: PluginType {
     
     public typealias NetworkActivityClosure = (change: NetworkActivityChangeType) -> ()
     let networkActivityClosure: NetworkActivityClosure
@@ -18,13 +18,12 @@ public final class NetworkActivityPlugin: Plugin {
     // MARK: Plugin
 
     /// Called by the provider as soon as the request is about to start
-    public func willSendRequest(request: MoyaRequest, target: MoyaTarget) {
+    public func willSendRequest(request: RequestType, target: TargetType) {
         networkActivityClosure(change: .Began)
     }
     
     /// Called by the provider as soon as a response arrives
-    public func didReceiveResponse(result: Result<Moya.Response, Moya.Error>, target: MoyaTarget) {
+    public func didReceiveResponse(result: Result<Moya.Response, Moya.Error>, target: TargetType) {
         networkActivityClosure(change: .Ended)
     }
 }
-